move Web3Connect to common
