+++ src/common/web3connect/providers/InjectedProvider.js
