+++ src/common/web3connect/providers/WalletConnectProvider.js
